pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-stream-replay"
version = "0.1.0"
description = "Modular VRS-to-stream replay frontend for incremental Gaussian SLAM backends"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "hydra-core",
  "omegaconf",
  "numpy",
  "opencv-python",
  "pyzmq",
  "rich",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
